requirements.txt
Django
django-filter==2.2.0
djangorestframework==3.11.1
django-cors-headers
django-tinymce
joblib==0.16.0
numpy==1.17.0
pandas==1.1.1
Pillow==6.1.0
psycopg2-binary==2.8.3
python-dateutil==2.8.1
pytz==2019.1
requests
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sqlparse==0.3.0
threadpoolctl==2.1.0
